feat(hooks): Finalize default initializer functionality
