+++ packages/hooks/test/object.test.ts
@@ -7,14 +7,8 @@ interface HookableObject {
   addOne (number: number): Promise<number>;
 }
 
-interface Dummy {
-  sayHi (name: string): Promise<string>;
-  addOne (number: number): Promise<number>;
-}
-
 describe('objectHooks', () => {
   let obj: HookableObject;
-  let DummyClass: new () => Dummy;
 
   beforeEach(() => {
     obj = {
@@ -23,16 +17,6 @@ describe('objectHooks', () => {
       async sayHi (name: string) {
         return `Hi ${name}`;
       },
-
-      async addOne (number: number) {
-        return number + 1;
-      }
-    };
-
-    DummyClass = class DummyClass implements Dummy {
-      async sayHi (name: string) {
-        return `Hi ${name}`;
-      }
 
       async addOne (number: number) {
         return number + 1;
@@ -128,64 +112,6 @@ describe('objectHooks', () => {
     }
   });
 
-  it('hooking object on class adds to the prototype', async () => {
-    hooks(DummyClass, {
-      sayHi: middleware([async (ctx: HookContext, next: NextFunction) => {
-        assert.deepStrictEqual(ctx, new DummyClass.prototype.sayHi.Context({
-          arguments: ['David'],
-          method: 'sayHi',
-          name: 'David',
-          self: instance
-        }));
-
-        await next();
-
-        ctx.result += '?';
-      }]).params('name'),
-
-      addOne: middleware([async (ctx: HookContext, next: NextFunction) => {
-        ctx.arguments[0] += 1;
-
-        await next();
-      }])
-    });
-
-    const instance = new DummyClass();
-
-    assert.strictEqual(await instance.sayHi('David'), 'Hi David?');
-    assert.strictEqual(await instance.addOne(1), 3);
-  });
-
-  it('works with inheritance', async () => {
-    hooks(DummyClass, {
-      sayHi: middleware([async (ctx: HookContext, next: NextFunction) => {
-        assert.deepStrictEqual(ctx, new (OtherDummy.prototype.sayHi as any).Context({
-          arguments: [ 'David' ],
-          method: 'sayHi',
-          self: instance
-        }));
-
-        await next();
-
-        ctx.result += '?';
-      }])
-    });
-
-    class OtherDummy extends DummyClass {}
-
-    hooks(OtherDummy, {
-      sayHi: middleware([async (ctx: HookContext, next: NextFunction) => {
-        await next();
-
-        ctx.result += '!';
-      }])
-    });
-
-    const instance = new OtherDummy();
-
-    assert.strictEqual(await instance.sayHi('David'), 'Hi David?!');
-  });
-
   it('works with object level hooks', async () => {
     hooks(obj, [
       async (ctx: HookContext, next: NextFunction) => {
@@ -204,51 +130,5 @@ describe('objectHooks', () => {
     });
 
     assert.equal(await obj.sayHi('Dave'), 'Hi Dave?!');
-  });
-
-  it('works with multiple context updaters', async () => {
-    hooks(DummyClass, {
-      sayHi: middleware([
-        async (ctx, next) => {
-          assert.equal(ctx.name, 'Dave');
-          assert.equal(ctx.gna, 42);
-          assert.equal(ctx.app, 'ok');
-
-          ctx.name = 'Changed';
-
-          await next();
-        }
-      ]).params('name')
-    });
-
-    class OtherDummy extends DummyClass {}
-
-    hooks(OtherDummy, {
-      sayHi: middleware([
-        async (ctx, next) => {
-          assert.equal(ctx.name, 'Dave');
-          assert.equal(ctx.gna, 42);
-          assert.equal(ctx.app, 'ok');
-
-          await next();
-        }
-      ]).props({ gna: 42 })
-    });
-
-    const instance = new OtherDummy();
-
-    hooks(instance, {
-      sayHi: middleware([
-        async (ctx, next) => {
-          assert.equal(ctx.name, 'Dave');
-          assert.equal(ctx.gna, 42);
-          assert.equal(ctx.app, 'ok');
-
-          await next();
-        }
-      ]).props({ app: 'ok' })
-    });
-
-    assert.equal(await instance.sayHi('Dave'), 'Hi Changed');
   });
 });
